+++ eat_all_fungus/lib/constValues/helperFunctions.dart
@@ -1,6 +1,5 @@
 import 'package:eat_all_fungus/controllers/profileController.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 int countAmountOfItems(List<String> input, String toCount) {
