+++ eat_all_fungus/lib/providers/streams/townStream.dart
@@ -8,7 +8,6 @@ import 'package:eat_all_fungus/models/world.dart';
 import 'package:eat_all_fungus/providers/streams/playerStream.dart';
 import 'package:eat_all_fungus/providers/streams/tileStream.dart';
 import 'package:eat_all_fungus/providers/streams/worldStream.dart';
-import 'package:eat_all_fungus/services/tileRepository.dart';
 import 'package:eat_all_fungus/services/townRepository.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
