+++ eat_all_fungus/lib/views/inGame/map/mapWidget.dart
@@ -4,9 +4,8 @@ import 'package:eat_all_fungus/providers/streams/playerStream.dart';
 import 'package:eat_all_fungus/providers/streams/tileStream.dart';
 import 'package:eat_all_fungus/providers/streams/worldStream.dart';
 import 'package:eat_all_fungus/views/various/loadings/loadingsWidget.dart';
-import 'package:eat_all_fungus/views/widgets/buttons/digButton.dart';
 import 'package:eat_all_fungus/views/widgets/buttons/mapButton.dart';
-import 'package:eat_all_fungus/views/widgets/items/inventory.dart';
+import 'package:eat_all_fungus/views/widgets/items/inventories/tileInventory.dart';
 import 'package:eat_all_fungus/views/widgets/mapView/mapSubWidgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
