+++ eat_all_fungus/lib/views/inGame/player/playerSubWidgets.dart
@@ -2,9 +2,10 @@ import 'package:eat_all_fungus/constValues/constValues.dart';
 import 'package:eat_all_fungus/controllers/townController.dart';
 import 'package:eat_all_fungus/providers/streams/playerStream.dart';
 import 'package:eat_all_fungus/providers/streams/tileStream.dart';
-import 'package:eat_all_fungus/views/inGame/overview/overviewSubWidgets.dart';
 import 'package:eat_all_fungus/views/widgets/buttons/townButton.dart';
-import 'package:eat_all_fungus/views/widgets/items/inventory.dart';
+import 'package:eat_all_fungus/views/widgets/constWidgets/panel.dart';
+import 'package:eat_all_fungus/views/widgets/items/inventories/tileInventory.dart';
+import 'package:eat_all_fungus/views/widgets/items/itemPanel.dart';
 import 'package:eat_all_fungus/views/widgets/items/townStash.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
