+++ eat_all_fungus/lib/views/widgets/buttons/digButton.dart
@@ -1,7 +1,6 @@
 import 'package:eat_all_fungus/constValues/constValues.dart';
 import 'package:eat_all_fungus/controllers/functionsController.dart';
 import 'package:eat_all_fungus/providers/streams/digTaskStream.dart';
-import 'package:eat_all_fungus/views/inGame/overview/overviewSubWidgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
