+++ eat_all_fungus/lib/views/widgets/items/townStash.dart
@@ -1,16 +1,16 @@
 import 'package:eat_all_fungus/constValues/constValues.dart';
-import 'package:eat_all_fungus/controllers/townController.dart';
 import 'package:eat_all_fungus/providers/streams/playerStream.dart';
 import 'package:eat_all_fungus/providers/streams/stashStream.dart';
 import 'package:eat_all_fungus/providers/streams/townStream.dart';
-import 'package:eat_all_fungus/views/inGame/overview/overviewSubWidgets.dart';
-import 'package:eat_all_fungus/views/widgets/items/inventory.dart';
+import 'package:eat_all_fungus/views/widgets/constWidgets/panel.dart';
+import 'package:eat_all_fungus/views/widgets/items/inventories/stashInventory.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class TownStash extends HookWidget {
-  const TownStash({Key? key}) : super(key: key);
+  final bool canWrap;
+  const TownStash({Key? key, this.canWrap = false}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -18,8 +18,8 @@ class TownStash extends HookWidget {
     final playerStream = useProvider(playerStreamProvider);
     if (townStream?.members.contains(playerStream?.id) ?? false) {
       final stashState = useProvider(stashStreamProvider);
-      final itemWidgetList =
-          buildTileInventoryList(tileInventory: stashState ?? []);
+      final itemWidgetList = buildStashInventoryList(
+          tileInventory: stashState ?? [], canTap: true);
       return Container(
         color: Colors.grey[colorIntensity],
         child: Center(
@@ -27,14 +27,38 @@ class TownStash extends HookWidget {
           padding: const EdgeInsets.all(8.0),
           child: Column(
             children: [
-              Text('Items on Tile:'),
+              Text('Items in Stash:'),
               Expanded(
-                child: ListView(
-                  scrollDirection: Axis.horizontal,
-                  shrinkWrap: true,
-                  physics: ClampingScrollPhysics(),
-                  children: itemWidgetList,
-                ),
+                child: canWrap
+                    ? Container(
+                        child: SingleChildScrollView(
+                          child: Wrap(
+                            children: [
+                              ...itemWidgetList
+                                  .map((e) => Padding(
+                                        padding: const EdgeInsets.all(8.0),
+                                        child: Container(
+                                          decoration: BoxDecoration(
+                                              borderRadius:
+                                                  BorderRadius.circular(15.0),
+                                              border: Border.all(
+                                                  color: Colors.grey)),
+                                          width: 100,
+                                          height: 100,
+                                          child: Panel(child: e),
+                                        ),
+                                      ))
+                                  .toList()
+                            ],
+                          ),
+                        ),
+                      )
+                    : ListView(
+                        scrollDirection: Axis.horizontal,
+                        shrinkWrap: true,
+                        physics: ClampingScrollPhysics(),
+                        children: itemWidgetList,
+                      ),
               )
             ],
           ),
